+++ docs/CHANGE_LOG.md
@@ -1,7 +1,9 @@
 # Heat.js - Change Log:
 
 ## Version 0.5.0:
-- 
+
+#### **General:**
+- Updated the projects main description.
 
 <br>