+++ docs/CHANGE_LOG.md
@@ -1,7 +1,9 @@
 # Heat.js - Change Log:
 
 ## Version 0.6.0:
-- 
+
+#### **Fixes:**
+- Fixed a fault that allowed date counts to go into negative values (when calling "removeDate()").
 
 <br>